updating the test matrix for legacy python (#56)
GitHub Actions recently changed the "ubuntu-latest" tag to invoke "ubuntu-22.04" instead of "ubuntu-20.04".  The 22.04 build and python actions no longer support the 2.7 and 3.6 versions, which caused many CIs to fail.  Support for 2.7 was eventually brought back, support for 3.6 may not have been.

This commit uses include/exclude rules to run 2.7 and 3.6 explicitly on the ubuntu-20.04 platform.  It also utilizes custom naming of tests, to include the platform+python+redis versions to better pinpoint any failures.